Extract Diagnostic into own mod
