+++ crates/mockiato-codegen/src/parse/method_decl.rs
@@ -1,7 +1,8 @@
 use super::check_option_is_none;
+use crate::diagnostic::DiagnosticBuilder;
 use crate::parse::method_inputs::MethodInputs;
+use crate::result::{merge_results, Error, Result};
 use crate::syn_ext::PathExt;
-use crate::{merge_results, DiagnosticBuilder, Error, Result};
 use proc_macro2::Span;
 use std::collections::HashSet;
 use syn::spanned::Spanned;
